fix bugs in ScopeNonExactPredicatesTransformer unapply
