+++ core/src/main/scala/fortress/modelfind/CompilationModelFinder.scala
@@ -91,90 +91,6 @@ with ModelFinderSettings {
         val instance = solverSession.get.solution
             .withoutDeclarations(compilerResult.get.skipForNextInterpretation)
 
-        def forceValueToBool(term: Value): Boolean = term match{
-            case Top => true
-            case Bottom => false
-            case _ => Errors.Internal.impossibleState("Tried to cast non-Top/Bottom Term to Boolean")
-        }
-        def boolToValue(b: Boolean): Value = if(b) Top else Bottom
-
-
-        def getFunctionValue(fnName: String, evaluatedArgs: Seq[Value]): Value = {
-            val funcDef = instance.functionDefinitions.filter(fd => fd.name == fnName).head
-            val formalArgs: Seq[Term] = for( item <- funcDef.argSortedVar ) yield item.variable
-            // transfer constants to domain elements, ex: p1 -> _@1P
-            val realArgs: Seq[Value] = for( item <- evaluatedArgs ) yield {
-                var temp: Value = item
-                // TODO: look here! @zxt
-//                for( a <- instance.constantInterpretations ) {
-//                    if(a._1.variable.name == temp.toString ) temp = a._2
-//                }
-                temp
-            }
-            val body = funcDef.body
-            Errors.Internal.precondition(evaluatedArgs.size == formalArgs.size, "Invalid input params.")
-            val argMap: Map[Term, Value] = formalArgs.zip(realArgs).toMap
-            val ret: Value = visitFunctionBody( body, argMap )
-            ret
-        }
-
-        def visitFunctionBody(term: Term, argMap: Map[Term, Value]): Value = term match {
-            case Top | Bottom | EnumValue(_) | DomainElement(_, _) |
-                 IntegerLiteral(_) | BitVectorLiteral(_, _) => term.asInstanceOf[Value]
-            case v @ Var(_) => argMap(v)
-            case Not(p) => boolToValue(!forceValueToBool(visitFunctionBody(p, argMap)))
-            case AndList(args) => boolToValue(args.forall(a => forceValueToBool(visitFunctionBody(a, argMap))))
-            case OrList(args) => boolToValue(args.exists(a => forceValueToBool(visitFunctionBody(a, argMap))))
-            case Distinct(args) => boolToValue(
-                args.size == args.map(a => visitFunctionBody(a, argMap)).distinct.size
-            )
-            case Implication(p, q) => boolToValue(
-                !forceValueToBool(visitFunctionBody(p, argMap)) || forceValueToBool(visitFunctionBody(q, argMap))
-            )
-            case Iff(p, q) => boolToValue(
-                forceValueToBool(visitFunctionBody(p, argMap)) == forceValueToBool(visitFunctionBody(q, argMap))
-            )
-            case Eq(l, r) => boolToValue(visitFunctionBody(l, argMap) == visitFunctionBody(r, argMap))
-            case IfThenElse(condition, ifTrue, ifFalse) => {
-                if(forceValueToBool(visitFunctionBody(condition, argMap))) {
-                    visitFunctionBody(ifTrue, argMap)
-                }
-                else {
-                    visitFunctionBody(ifFalse, argMap)
-                }
-            }
-            case App(fname, args) => getFunctionValue(fname, args.map(arg => visitFunctionBody(arg, argMap)))
-            case BuiltinApp(fn, args) => evaluateBuiltIn(fn, args.map(arg => visitFunctionBody(arg, argMap)))
-            case _ => {
-                println("Error: get function value failed.")
-                null
-            }
-        }
-
-        // Given a builtin function and its arguments, run it through a throwaway Z3 solver for the result
-        // (to avoid having to implement every function manually on our end)
-        def evaluateBuiltIn(fn: BuiltinFunction, evalArgs: Seq[Value]): Value = {
-            val evalResult: Var = Var("!VERIFY_INTERPRETATION_RES")
-            val evalResultAnnotated: AnnotatedVar = fn match {
-                case IntPlus | IntNeg | IntSub | IntMult | IntDiv | IntMod => evalResult of Sort.Int
-                case BvPlus | BvNeg | BvSub | BvMult | BvSignedDiv | BvSignedRem | BvSignedMod =>
-                    evalResult of Sort.BitVector(evalArgs.head.asInstanceOf[BitVectorLiteral].bitwidth);
-                case IntLE | IntLT | IntGE | IntGT |
-                     BvSignedLE | BvSignedLT | BvSignedGE | BvSignedGT => evalResult of Sort.Bool
-                case _ => throw new scala.NotImplementedError("Builtin function not accounted for")
-            }
-            val theory: Theory = Theory.empty
-                    .withConstant(evalResultAnnotated)
-                    .withAxiom(evalResult === BuiltinApp(fn, evalArgs))
-
-            val solver = new Z3IncSolver
-            solver.setTheory(theory)
-            solver.solve(Milliseconds(1000))
-            val solvedInstance = solver.solution
-            solver.close()
-            solvedInstance.constantInterpretations(evalResultAnnotated)
-        }
-
         val newInstance: Interpretation = {
 
             def scopes: Map[Sort, Int] = for((sort, seq) <- instance.sortInterpretations) yield (sort -> seq.size)
@@ -183,7 +99,7 @@ with ModelFinderSettings {
                 for(f <- theory.signature.functionDeclarations)
                     yield f -> {
                         for (argList <- ArgumentListGenerator.generate(f, scopes, Some(instance.sortInterpretations)))
-                            yield argList -> getFunctionValue(f.name, argList)
+                            yield argList -> instance.getFunctionValue(f.name, argList)
                         }.toMap
 
             }.toMap
