+++ core/src/main/scala/fortress/transformers/ScopeNonExactPredicatesTransformer.scala
@@ -51,13 +51,15 @@ class ScopeNonExactPredicatesTransformer extends ProblemStateTransformer {
             val unapply: Interpretation => Interpretation = {
                 interp => {
                     var resultInterp = interp
-                    for (pred <- scopeNonExactPreds) {  // for each subtype predicate funcDecl
-                        val sort: Sort = pred.argSorts.head // get argument sort
-                        // filter: get preds whose result sort is Top(True)
-                        val domainElements: Seq[Value] = interp.functionInterpretations(pred).toSeq.filter(pred => pred._2 == Top).map(_._1.head)
-                        resultInterp = resultInterp.updateSortInterpretations(sort, domainElements)
+                    val sortInterp: Map[Sort, Seq[Value]] = interp.sortInterpretations
+                    for(pred <- scopeNonExactPreds) {
+                        val sort: Sort = pred.argSorts.head
+                        val Interp: Seq[Value] = sortInterp(sort).filter(value => {
+                            interp.getFunctionValue(pred.name, Seq(value)) == Top
+                        })
+                        resultInterp = resultInterp.updateSortInterpretations(sort, Interp)
                     }
-                    resultInterp
+                    resultInterp.withoutFunctions(scopeNonExactPreds)
                 }
             }
